# specs/001-create-a-drag/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Teacher Aide Scheduler API
  version: 1.0.0
  description: REST API for drag-and-drop timetable scheduling system
servers:
  - url: http://localhost:5000/api
    description: Local development server
paths:
...
  /aides:
    get:
      summary: List all teacher aides
      tags: [Teacher Aides]
      responses:
        '200':
          description: List of teacher aides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeacherAide'
    post:
      summary: Create a new teacher aide
      tags: [Teacher Aides]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, colour_hex]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                qualifications:
                  type: string
                colour_hex:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '201':
          description: Aide created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherAide'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /aides/{id}:
    get:
      summary: Get aide by ID
      tags: [Teacher Aides]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Aide details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherAide'
        '404':
          description: Aide not found
    put:
      summary: Update aide
      tags: [Teacher Aides]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                qualifications:
                  type: string
                colour_hex:
                  type: string
      responses:
        '200':
          description: Aide updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherAide'
    delete:
      summary: Delete aide
      tags: [Teacher Aides]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Aide deleted
        '404':
          description: Aide not found
  
  # Availability
...
  /assignments:
    get:
      summary: List assignments
      tags: [Assignments]
      parameters:
        - name: week
          in: query
          schema:
            type: string
            pattern: '^\d{4}-W\d{2}$'
            example: "2025-W01"
        - name: aide_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [UNASSIGNED, ASSIGNED, IN_PROGRESS, COMPLETE]
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
    post:
      summary: Create single assignment
      tags: [Assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '409':
          description: Conflict detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
  /assignments/weekly-matrix:
    get:
      summary: Get structured weekly matrix for UI
      tags: [Assignments]
      parameters:
        - name: week
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-W\d{2}$'
            example: "2025-W01"
      responses:
        '200':
          description: Weekly matrix data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyMatrix'
  /assignments/batch:
    post:
      summary: Batch create assignments (recurring task multi-day)
      tags: [Assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_id, aide_id, dates]
              properties:
                task_id:
                  type: integer
                aide_id:
                  type: integer
                dates:
                  type: array
                  items:
                    type: string
                    format: date
      responses:
        '201':
          description: Assignments created
          content:
            application/json:
              schema:
                type: object
                properties:
                  assignments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Assignment'
                  conflicts:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConflictError'
...
  /assignments/{id}:
    get:
      summary: Get assignment by ID
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Assignment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
    put:
      summary: Update assignment (with version check for optimistic locking)
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                aide_id:
                  type: integer
                  nullable: true
                start_time:
                  type: string
                end_time:
                  type: string
                status:
                  type: string
                  enum: [UNASSIGNED, ASSIGNED, IN_PROGRESS, COMPLETE]
                version:
                  type: string
                  format: date-time
                  description: Client's version timestamp for optimistic locking
      responses:
        '200':
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '409':
          description: Conflict (version mismatch or time collision)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
    delete:
      summary: Delete assignment
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Assignment deleted
  
  # Absences
...
components:
  schemas:
    TeacherAide:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        qualifications:
          type: string
        colour_hex:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        task_id:
          type: integer
        aide_id:
          type: integer
          nullable: true
        date:
          type: string
          format: date
        start_time:
          type: string
        end_time:
          type: string
        status:
          type: string
          enum: [UNASSIGNED, ASSIGNED, IN_PROGRESS, COMPLETE]
        version:
          type: string
          format: date-time
        task_title:
          type: string
        task_category:
          type: string
        classroom:
          type: string
          nullable: true
    AssignmentCreate:
      type: object
      required: [task_id, date, start_time, end_time]
      properties:
        task_id:
          type: integer
        aide_id:
          type: integer
          nullable: true
        date:
          type: string
          format: date
        start_time:
          type: string
        end_time:
          type: string
    Absence:
      type: object
      properties:
        id:
          type: integer
        aide_id:
          type: integer
        date:
          type: string
          format: date
        reason:
          type: string
        created_at:
          type: string
          format: date-time
...
    WeeklyMatrix:
      type: object
      properties:
        week:
          type: string
          example: "2025-W01"
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        time_slots:
          type: array
          items:
            type: string
          example: ["08:00", "08:30", "09:00", "09:30", "10:00", "10:30", "11:00", "11:30", "12:00", "12:30", "13:00", "13:30", "14:00", "14:30", "15:00", "15:30"]
        days:
          type: array
          items:
            type: string
          example: ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday"]
        aides:
          type: array
          items:
            $ref: '#/components/schemas/TeacherAide'
        assignments:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Assignment'
          example:
            "1_Monday_09:00":
              assignment_id: 123
              task_id: 456
              task_title: "Morning Playground Duty"
              task_category: "PLAYGROUND"
              start_time: "09:00"
              end_time: "10:00"
              status: "ASSIGNED"
        absences:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Absence'
          example:
            "1_Monday":
              absence_id: 789
              reason: "Sick leave"
              date: "2025-01-06"
    ConflictError:
      type: object
      properties:
        error:
          type: string
          example: "Assignment conflict"
        conflict_type:
          type: string
          enum: [time_collision, version_mismatch, partial_overlap]
        conflicting_assignment:
          type: object
          properties:
            id:
              type: integer
            task_title:
              type: string
            start_time:
              type: string
            end_time:
              type: string
            aide_name:
              type: string
        suggestion:
          type: string
          enum: [replace, shorten_first, cancel]
        current_version:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string